requirements.txt
black==22.8.0
pylint==2.15.3
pytest==7.1.3
pytest-cov==4.0.0
python-dotenv==0.21.0
requests==2.28.1